backend updates

+++ server/cinema-app-back/src/main/java/com/cinema/web/controller/MovieController.java
@@ -10,12 +10,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 
 @RestController
 @RequestMapping(value = "/api/movies", produces = MediaType.APPLICATION_JSON_VALUE)
+@Validated
 public class MovieController {
 
     @Autowired
@@ -29,33 +32,32 @@ public class MovieController {
 
     @Autowired
     private MovieDTOToMovieUpdate toMovieUpdate;
+/*
+    @GetMapping
+    public ResponseEntity<List<MovieDTO>> getAll() {
 
-    /*
-     * @GetMapping
-     * public ResponseEntity<List<MovieDTO>> getAll() {
-     * 
-     * List<Movie> movies = movieService.findAll();
-     * 
-     * return new ResponseEntity<>(toDTO.convertAll(movies), HttpStatus.OK);
-     * }
-     */
-    // @PreAuthorize("hasRole('ROLE_ADMIN')")
+        List<Movie> movies = movieService.findAll();
+
+        return new ResponseEntity<>(toDTO.convertAll(movies), HttpStatus.OK);
+    }
+   */ 
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping
     public ResponseEntity<List<MovieDTO>> getList(
-            @RequestParam(required = false) String name,
-            @RequestParam(required = false) Integer durationMin,
-            @RequestParam(required = false) Integer durationMax,
-            @RequestParam(required = false) String country,
-            @RequestParam(required = false) String distributor,
-            @RequestParam(required = false) Integer yearMin,
-            @RequestParam(required = false) Integer yearMax,
-            @RequestParam(required = false) String sortBy,
-            @RequestParam(required = false) String sortAscOrDesc) {
-
-        List<Movie> movies = movieService.findByParameters(name, durationMin, durationMax, country, distributor,
-                yearMin, yearMax, sortBy, sortAscOrDesc);
-        return new ResponseEntity<>(toDTO.convertAll(movies), HttpStatus.OK);
-
+    		@RequestParam(required = false) String name,
+    		@RequestParam(required = false) Integer durationMin,
+    		@RequestParam(required = false) Integer durationMax, 
+    		@RequestParam(required = false) String country, 
+    		@RequestParam(required = false) String distributor,
+    		@RequestParam(required = false) Integer yearMin,
+    		@RequestParam(required = false) Integer yearMax,
+    		@RequestParam(required = false) String sortBy,
+    		@RequestParam(required = false) String sortAscOrDesc
+    		){
+        	
+    	List<Movie> movies = movieService.findByParameters(name, durationMin, durationMax, country, distributor,yearMin, yearMax, sortBy, sortAscOrDesc);
+    	movies.forEach(m -> System.out.println(m));
+    	return new ResponseEntity<>(toDTO.convertAll(movies), HttpStatus.OK); 
     }
 
     @GetMapping(value = "/{id}")
@@ -63,49 +65,47 @@ public class MovieController {
         Movie movie = movieService.findById(id);
         if (movie != null) {
             return new ResponseEntity<>(toDTO.convert(movie), HttpStatus.OK);
-
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
-
-    // @PreAuthorize("hasRole('ROLE_ADMIN')")
+    //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<MovieDTO> create(@RequestBody MovieDTO movieDTO) {
+    public ResponseEntity<MovieDTO> create(@Valid @RequestBody MovieDTO movieDTO) {
         Movie newMovie = movieService.save(toMovieNew.convert(movieDTO));
 
         return new ResponseEntity<>(toDTO.convert(newMovie), HttpStatus.CREATED);
     }
-
-    // @PreAuthorize("hasRole('ROLE_ADMIN')")
+  //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<MovieDTO> update(@PathVariable Long id, @Valid @RequestBody MovieDTO movieDTO) {
         if (!id.equals(movieDTO.getId())) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        Movie movie = movieService.findById(id);
+        Movie movie =  movieService.findById(id);
         if (movie == null) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        if (movieDTO.getVersion() != movie.getVersion()) {
-
-            return new ResponseEntity<>(toDTO.convert(movie), HttpStatus.BAD_REQUEST);
+        if(movieDTO.getVersion() != movie.getVersion()) {
+        	return new ResponseEntity<>(toDTO.convert(movie), HttpStatus.BAD_REQUEST);
         }
         Movie movieUpdate = toMovieUpdate.convert(movieDTO);
         Movie savedMovie = movieService.update(movieUpdate);
 
         return new ResponseEntity<>(toDTO.convert(savedMovie), HttpStatus.OK);
     }
-
-    // @PreAuthorize("hasRole('ROLE_ADMIN')")
+  //@PreAuthorize("hasRole('ROLE_ADMIN')")
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+    	 if(movieService.findById(id) == null)
+    		 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    	
         Movie deleted = movieService.delete(id);
 
         if (deleted != null) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
     /*
